Rework BackedUpFileListCommand CloudFilesRegistry.backedUpList CloudFilesRegistryTest

+++ app/lib/server/CloudFilesRegistry.scala
@@ -1,5 +1,7 @@
 package lib.server
 
+import java.io.File.{separatorChar => SystemSlashChar}
+
 import better.files.File
 import cats.data.NonEmptyList
 import cats.syntax.all._
@@ -7,35 +9,21 @@ import com.typesafe.scalalogging.StrictLogging
 import controllers.WsApiController
 import io.circe.Json
 import io.circe.generic.extras.auto._
+import io.circe.syntax._
 import javax.inject.Inject
 import lib.App
-import lib.App._
-import lib.client.clientapi.FileTreeNode.{Directory, RegularFile}
+import lib.App.{StringOps, _}
+import lib.client.clientapi.FileTreeNode.Directory
 import lib.client.clientapi.{FileTreeNode, Version}
 import lib.db.{DbFile, FilesDao}
+import lib.server.CloudFilesRegistry._
 import lib.server.serverapi.{RemoteFile, RemoteFileVersion}
+import org.apache.commons.lang3.StringUtils
 
 class CloudFilesRegistry @Inject()(wsApiController: WsApiController, dao: FilesDao) extends StrictLogging {
 
   def updateFile(file: File, remoteFile: RemoteFile): Result[Unit] = {
     dao.update(file, remoteFile)
-  }
-
-  def reportBackedUpFilesList: Result[Unit] = {
-    def sendWsUpdate(json: Json): Result[Unit] = {
-      wsApiController.send(controllers.WsMessage(`type` = "backedUpFilesUpdate", data = json), ignoreFailure = true)
-    }
-
-    ???
-
-    //    for {
-    //      files <- dao.listAll()
-    //      fileTrees = FileTree.fromRemoteFiles(files.map(_.remoteFile))
-    //      json = fileTrees.toJson
-    //      _ <- sendWsUpdate(json)
-    //    } yield {
-    //      ()
-    //    }
   }
 
   def versions(file: File): Result[Option[NonEmptyList[RemoteFileVersion]]] = {
@@ -50,7 +38,7 @@ class CloudFilesRegistry @Inject()(wsApiController: WsApiController, dao: FilesD
     App.leaveBreadcrumb("Removing file", Map("path" -> file.originalName))
 
     dao.delete(file) >>
-      reportBackedUpFilesList
+      reportDirUpdate(File(file.originalName).parent)
   }
 
   def removeFileVersion(file: File, fileVersion: RemoteFileVersion): Result[Unit] = {
@@ -60,11 +48,12 @@ class CloudFilesRegistry @Inject()(wsApiController: WsApiController, dao: FilesD
       case Some(f) =>
         val restOfVersions = f.versions.filterNot(_ == fileVersion)
 
-        (if (restOfVersions.nonEmpty) {
-           updateFile(file, f.copy(versions = restOfVersions))
-         } else {
-           removeFile(f)
-         }) >> reportBackedUpFilesList
+        if (restOfVersions.nonEmpty) {
+          updateFile(file, f.copy(versions = restOfVersions)) >>
+            reportVersionsUpdate(f, restOfVersions)
+        } else {
+          removeFile(f)
+        }
 
       case None =>
         App.leaveBreadcrumb("Didn't remove file version because it couldn't be found")
@@ -72,7 +61,29 @@ class CloudFilesRegistry @Inject()(wsApiController: WsApiController, dao: FilesD
     }
   }
 
-  def list(prefix: Option[String]): Result[List[FileTreeNode]] = {
+  private def reportDirUpdate(parent: File): Result[Unit] = {
+    listFiles(prefix = Some(parent.pathAsString)).flatMap { content =>
+      wsApiController.send(
+        `type` = "removedFile",
+        data = parseUnsafe {
+          s"""{ "parent": ${parent.pathAsString.asJson}, "files": ${content.map(_.toJson).asJson}}"""
+        },
+        ignoreFailure = false
+      )
+    }
+  }
+
+  private def reportVersionsUpdate(f: RemoteFile, versions: List[RemoteFileVersion]): Result[Unit] = {
+    wsApiController.send(
+      `type` = "removedFileVersion",
+      data = parseUnsafe {
+        s"""{ "path": ${f.originalName.asJson}, "versions": ${versions.map(Version(f.originalName, _).toJson).asJson}}"""
+      },
+      ignoreFailure = false
+    )
+  }
+
+  def listFiles(prefix: Option[String]): Result[List[FileTreeNode]] = {
     dao.listAll(prefix.getOrElse("")).map { files =>
       if (files.nonEmpty) {
         backedUpList(prefix, files)
@@ -81,24 +92,53 @@ class CloudFilesRegistry @Inject()(wsApiController: WsApiController, dao: FilesD
       }
     }
   }
+}
 
-  private[server] def backedUpList(prefix: Option[String], files: List[DbFile]): List[FileTreeNode] = {
-    //noinspection MapGetOrElseBoolean
-    def isFile(dbFile: DbFile): Boolean = {
-      prefix
-        .map(p => dbFile.path.stripPrefix(p).replace("\\", "/").contains("/"))
-        .getOrElse(false)
-    }
-
-    files.map { dbFile =>
-      import dbFile._
-
-      if (isFile(dbFile)) {
-        RegularFile(File(path), NonEmptyList.fromList(remoteFile.versions))
-      } else {
-        Directory(path, prefix)
+object CloudFilesRegistry {
+  private[server] def backedUpList(prefix: Option[String], files: List[DbFile], systemSlash: Char = SystemSlashChar): List[FileTreeNode] = {
+    val trailedPrefix = prefix
+      .map { pr =>
+        if (!pr.endsWith(systemSlash.toString)) {
+          pr + systemSlash
+        } else pr
       }
-    }.distinct
+
+    val filteredFiles = trailedPrefix.map(p => files.filter(_.path.startsWith(p))).getOrElse(files)
+
+    val uniqueNames = filteredFiles
+      .map { dbFile =>
+        import dbFile._
+        trailedPrefix
+          .map(_.fixPath)
+          .map(path.fixPath.stripPrefix)
+          .getOrElse(path.fixPath)
+          .split("/")
+          .headOption
+          .filter(StringUtils.isNotBlank)
+          .getOrElse("/")
+      }
+      .distinct
+      .sorted
+
+    uniqueNames.map { name =>
+      val fullPath = (trailedPrefix.getOrElse("") + name).unfixPath
+
+      filteredFiles.find(_.path == fullPath) match {
+        case Some(dbFile) =>
+          FileTreeNode.RegularFile(
+            path = fullPath,
+            name = name,
+            versions = NonEmptyList.fromList(dbFile.remoteFile.versions.map(Version(fullPath, _)))
+          )
+
+        case None =>
+          Directory(
+            path = fullPath,
+            name = name,
+            children = None
+          )
+      }
+    }
   }
 }
